게임 플레이 화면 연결
